Move should_link and message_format to separate functions
