+++ src/main.rs
@@ -1,5 +1,8 @@
-use cargo_3ds::commands::{Cargo, CargoCommand};
-use cargo_3ds::{build_3dsx, build_elf, build_smdh, check_rust_version, get_metadata, link};
+use cargo_3ds::commands::Cargo;
+use cargo_3ds::{
+    build_3dsx, build_elf, build_smdh, check_rust_version, get_message_format, get_metadata,
+    get_should_link, link,
+};
 use clap::Parser;
 use std::process;
 
@@ -8,35 +11,8 @@ fn main() {
 
     let Cargo::Input(mut input) = Cargo::parse();
 
-    let should_link = input.cmd == CargoCommand::Build
-        || (input.cmd == CargoCommand::Test
-            && if input.cargo_opts.contains(&"--no-run".to_string()) {
-                false
-            } else {
-                input.cargo_opts.push("--no-run".to_string());
-                true
-            });
-
-    let message_format = if let Some(pos) = input
-        .cargo_opts
-        .iter()
-        .position(|s| s.starts_with("--message-format"))
-    {
-        let arg = input.cargo_opts.remove(pos);
-        let format = if let Some((_, format)) = arg.split_once('=') {
-            format.to_string()
-        } else {
-            input.cargo_opts.remove(pos)
-        };
-        if !format.starts_with("json") {
-            eprintln!("error: non-JSON `message-format` is not supported");
-            process::exit(1);
-        } else {
-            format
-        }
-    } else {
-        "json-render-diagnostics".to_string()
-    };
+    let should_link = get_should_link(&mut input);
+    let message_format = get_message_format(&mut input);
 
     let (status, messages) = build_elf(input.cmd, &message_format, &input.cargo_opts);
